Sepere los archivos java en diferentes partes por modularidad,hilo 1 lista con ruta absoluta  y agrege fecha al terminar cada hilo
